src/third_party/blink/web_tests/external/wpt/focus/focus-restoration-in-different-site-iframes-expected.txt

This is a testharness.js-based test.
FAIL Check result assert_equals: Check log expected "outerlog:log:bodyfocus,inputfocus,inputblur,bodyblur,bodyfocus,inputfocus," but got "outerlog:log:"
Harness: the test ran to completion.
